refactor(component): ♻️ Update component config
