--- packages/components/src/Button/interface.ts
@@ -1,3 +1,0 @@
-import type { ButtonProps } from 'naive-ui'
-
-export type UButtonProps = ButtonProps
