--- packages/components/src/Navigators/List.css
@@ -1,8 +1,0 @@
-/** Navigator */
-.navigator {
-  @apply text-base;
-
-  &-title {
-    @apply mb-2 p-4 text-gray-400 text-sm border-b;
-  }
-}
