--- packages/components/src/SimpleSearch/index.css
@@ -1,1 +1,0 @@
-/** SimpleSearch */
