--- packages/components/src/SimpleTable/index.css
@@ -1,1 +1,0 @@
-/** SimpleTable */
