+++ packages/components/src/UInput/Input.tsx
@@ -1,10 +1,11 @@
+import type { InputProps } from 'naive-ui'
 import { NInput } from 'naive-ui'
 import { defineComponent } from 'vue'
-import type { CInputProps } from './interface'
 import './styles/input.css'
+export type UInputProps = InputProps
 
-export default defineComponent<CInputProps>({
-  name: 'Input',
+export default defineComponent<UInputProps>({
+  name: 'UInput',
   setup(props) {
     return () => {
       return <NInput {...props} />
