+++ packages/components/src/UInput/index.ts
@@ -1,8 +1,6 @@
 export { default as Input } from './Input'
 export { default as InputGroup } from './InputGroup'
 export { default as InputGroupLabel } from './InputGroupLabel'
-export type {
-  CInputGroupLabelProps as InputGroupLabelProps,
-  CInputGroupProps as InputGroupProps,
-  CInputProps as InputProps
-} from './interface'
+export type { UInputProps } from './Input'
+export type { UInputGroupProps } from './InputGroup'
+export type { UInputGroupLabelProps } from './InputGroupLabel'
