+++ packages/components/src/index.ts
@@ -2,7 +2,5 @@
 import 'virtual:windi-components.css'
 import 'virtual:windi-utilities.css'
 
-export { default as UButton } from './Button'
-export type { UButtonProps } from './Button/interface'
-export { default as Navigators } from './Navigators'
-export { default as SimpleTable } from './SimpleTable'
+export { default as UButton } from './UButton'
+export * from './UInput'
