--- packages/components/src/input/interface.ts
@@ -1,5 +1,0 @@
-import type { InputGroupLabelProps, InputGroupProps, InputProps } from 'naive-ui'
-
-export type CInputProps = InputProps
-export type CInputGroupProps = InputGroupProps
-export type CInputGroupLabelProps = InputGroupLabelProps
